Fixed couple of issues
Issue: There was a issue prevented adding same item to the grocerylist
twice. If attemped, the item was also added to the itemlist, which
caused duplicates in the itemlist. This was obviously unwanted bahavior.
This was fixed by creating a database model for the association table.
Earlier the assosiation was done with simple SQlalchemy "table". That
caused the issue